refactor: move auth routes to own module
