+++ apps/api/src/account/api.ts
@@ -1,5 +1,5 @@
 import { GetUserDetailsResponseData } from 'spotify-analytics-types';
-import { ResourceNotFoundError, NotImplementedError } from 'spotify-analytics-errors';
+import { ResourceNotFoundError } from 'spotify-analytics-errors';
 
 import { makeLogger } from '../logger';
 import prisma from '../prismaClient';
@@ -37,20 +37,4 @@ export async function getUserDetails(userId: number): Promise<GetUserDetailsResp
 
     log.info(results, `(${getUserDetails.name}) - results`);
     return results;
-}
-
-export async function webFlowLogin(provider: string) {
-    log.info({ provider }, `(${webFlowLogin.name})`);
-    throw new NotImplementedError(webFlowLogin.name + ' not implemented');
-    return;
-}
-
-export async function login(credentials: { username: string; password: string }) {
-    log.info({ credentials }, `(${login.name})`);
-    throw new NotImplementedError(login.name + ' not implemented');
-}
-
-export async function logout(session: string) {
-    log.info({ session }, `(${logout.name})`);
-    throw new NotImplementedError(logout.name + ' not implemented');
 }
