+++ apps/api/src/account/routes.ts
@@ -1,11 +1,9 @@
 import expressPromiseRouter from 'express-promise-router';
-import { getUserDetailsHandler, loginHandler, logoutHandler } from './handlers';
+import { getUserDetailsHandler } from './handlers';
 
 function init() {
     const router = expressPromiseRouter();
     router.get('/me', getUserDetailsHandler);
-    router.post('/login', loginHandler);
-    router.post('/logout', logoutHandler);
     return router;
 }
